Surface Android platform trust rejection reason via NetEventSource
When the Android platform trust manager rejects the certificate chain, it
only signals a boolean verdict; the underlying reason is lost. Capture the
Java CertificateException text and pass it through JNI to the managed
callback, which logs it via NetEventSource when a listener is attached.

The redundant chainTrustedByPlatform boolean is dropped: a non-null error
string now encodes rejection, and the string is only marshalled inside the
NetEventSource.Log.IsEnabled() guard to avoid allocations when tracing is off.

+++ src/libraries/Common/src/Interop/Android/System.Security.Cryptography.Native.Android/Interop.Ssl.cs
@@ -62,7 +62,7 @@ internal static partial class Interop
 
         [LibraryImport(Interop.Libraries.AndroidCryptoNative, EntryPoint = "AndroidCryptoNative_RegisterRemoteCertificateValidationCallback")]
         internal static unsafe partial void RegisterRemoteCertificateValidationCallback(
-            delegate* unmanaged<IntPtr, int, bool> verifyRemoteCertificate);
+            delegate* unmanaged<IntPtr, IntPtr, bool> verifyRemoteCertificate);
 
         [LibraryImport(Interop.Libraries.AndroidCryptoNative, EntryPoint = "AndroidCryptoNative_SSLStreamInitialize")]
         private static unsafe partial int SSLStreamInitializeImpl(
